removed duplicates, simplified files, added feedback
